Moved views and small alteration
