+++ frontend/src/main.ts
@@ -3,7 +3,7 @@ import { createPinia } from 'pinia'
 import App from './App.vue'
 import vuetify from './plugins/vuetify'
 import router from './router'
-import './styles.css'
+import './styles/styles.css'
 
 const app = createApp(App)
 const pinia = createPinia()
