+++ frontend/src/router/index.ts
@@ -5,12 +5,12 @@ const routes = [
   {
     path: '/',
     name: 'Home',
-    component: () => import('@/views/HomeView.vue')
+    component: () => import('@/views/index/HomeView.vue')
   },
   {
     path: '/login',
     name: 'Login',
-    component: () => import('@/views/LoginView.vue'),
+    component: () => import('@/views/shared/LoginView.vue'),
     meta: { requiresGuest: true }
   },
   {
