+++ frontend/src/styles/styles.css
